Using adaptation kludge in lookup of correct fn to call for oneLineDescription
We're taking the oneLineDescription out of playtools.query (formerly
goonmill.query) and eventually making it possible to register, with a plugin,
a RuleFact class and some associated formatters.  Step one is to start moving
this code out of goonmill.  Goonmill in the future will register a formatter
for Monster, and playtools will contain a sample plugin which registers a
formatter for Spell.  _formattingkludge contains a simple lookup mechanism to
call the right one.  A real implementation of this will in the future reside
in Playtools.